LASB-2105 Restructuring and extra tests
