+++ crime-applications-adaptor/src/main/java/uk/gov/justice/laa/crime/applications/adaptor/mapper/crimeapply/CrimeApplyMapper.java
@@ -3,8 +3,10 @@ package uk.gov.justice.laa.crime.applications.adaptor.mapper.crimeapply;
 import org.apache.commons.lang3.EnumUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
-import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.*;
-import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.ClientDetails;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.CrimeApplication;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.InterestOfJustice;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.MagistrateCourt;
+import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.Supplier;
 import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.MaatApplication;
 import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.general.Ioj;
 import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.general.Provider;
@@ -14,39 +16,34 @@ import java.util.Collections;
 import java.util.List;
 
 /**
- * The responsibility of this class is to encapsulate the required logic to map from a
- * Criminal Applications Datastore response
+ * The responsibility of this class is to map from a
+ * "Criminal Applications Datastore" response
  * to a
- * Crime Applications Adaptor MAAT Application (which should be structured like a MAAT ApplicationDTO)
+ * "Crime Applications Adaptor" MAAT Application (which should be structured like a MAAT ApplicationDTO)
  */
 @Component
 public class CrimeApplyMapper {
 
     private static final String STATUS_REASON_DEFAULT_CURRENT = "current";
     private static final boolean COURT_CUSTODY_DEFAULT_FALSE = false;
-    private static final String PARTNER_CONTRARY_INTEREST_CODE_DEFAULT_NOCON = "NOCON";
-    private static final String APPLICANT_GENDER_DEFAULT_PREFER_NOT_TO_SAY = "Prefer not to say";
-    private static final boolean APPLICANT_HAS_PARTNER_DEFAULT_FALSE = false;
 
     private final CaseDetailsMapper caseDetailsMapper;
-    private final OffenceClassMapper offenceClassMapper;
+    private final ApplicantMapper applicantMapper;
     private final PassportedMapper passportedMapper;
 
     public CrimeApplyMapper() {
         caseDetailsMapper = new CaseDetailsMapper();
-        offenceClassMapper = new OffenceClassMapper();
+        applicantMapper = new ApplicantMapper();
         passportedMapper = new PassportedMapper();
     }
 
     public CrimeApplication mapToCrimeApplication(MaatApplication crimeApplyResponse) {
-
         CrimeApplication crimeApplication = new CrimeApplication();
+
         crimeApplication.setStatusReason(STATUS_REASON_DEFAULT_CURRENT);
         crimeApplication.setSolicitorName(mapSolicitorName(crimeApplyResponse.getProviderDetails()));
         crimeApplication.setApplicationType(crimeApplyResponse.getApplicationType());
-
         crimeApplication.setCaseDetails(caseDetailsMapper.map(crimeApplyResponse.getCaseDetails()));
-
         crimeApplication.setMagsCourt(mapMagistrateCourt(crimeApplyResponse.getCaseDetails()));
         crimeApplication.setInterestsOfJustice(mapInterestsOfJustice(crimeApplyResponse.getInterestsOfJustice()));
         crimeApplication.setUsn(mapUsn(crimeApplyResponse));
@@ -55,7 +52,7 @@ public class CrimeApplyMapper {
         crimeApplication.setDateCreated(crimeApplyResponse.getSubmittedAt());
         crimeApplication.setDateStamp(crimeApplyResponse.getDateStamp());
         crimeApplication.setHearingDate(mapHearingDate(crimeApplyResponse.getCaseDetails()));
-        crimeApplication.setApplicant(mapApplicant(crimeApplyResponse));
+        crimeApplication.setApplicant(applicantMapper.map(crimeApplyResponse.getClientDetails()));
         crimeApplication.setSupplier(mapSupplier(crimeApplyResponse.getProviderDetails()));
         crimeApplication.setPassported(passportedMapper.map(crimeApplyResponse));
 
@@ -95,21 +92,22 @@ public class CrimeApplyMapper {
     }
 
     private MagistrateCourt mapMagistrateCourt(uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.CaseDetails crimeApplyCaseDetails) {
+        MagistrateCourt magistrateCourt = new MagistrateCourt();
+
         if (crimeApplyCaseDetails == null) {
-            return null;
+            return magistrateCourt;
         }
 
-        MagistrateCourt magistrateCourt = new MagistrateCourt();
         magistrateCourt.setCourt(crimeApplyCaseDetails.getHearingCourtName());
         return magistrateCourt;
     }
 
     private Supplier mapSupplier(Provider crimeApplyProviderDetails) {
-        if (crimeApplyProviderDetails == null) {
-            return null;
-        }
-
         Supplier supplier = new Supplier();
+
+        if (crimeApplyProviderDetails == null) {
+            return supplier;
+        }
 
         supplier.setOfficeCode(crimeApplyProviderDetails.getOfficeCode());
         supplier.setEmail(crimeApplyProviderDetails.getProviderEmail());
@@ -128,88 +126,5 @@ public class CrimeApplyMapper {
         List<String> firstAndLastName = List.of(crimeApplyProviderDetails.getLegalRepFirstName(),
                 crimeApplyProviderDetails.getLegalRepLastName());
         return String.join(StringUtils.SPACE, firstAndLastName);
-    }
-
-    private Applicant mapApplicant(MaatApplication crimeApplyResponse) {
-        Applicant applicant = new Applicant();
-
-        ClientDetails crimeApplyClientDetails = crimeApplyResponse.getClientDetails();
-        if (crimeApplyClientDetails == null) {
-            return null;
-        }
-
-        uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.Applicant crimeApplyApplicant =
-                crimeApplyClientDetails.getApplicant();
-        if (crimeApplyApplicant == null) {
-            return null;
-        }
-
-        applicant.setFirstName(crimeApplyApplicant.getFirstName());
-        applicant.setOtherNames(crimeApplyApplicant.getOtherNames());
-        applicant.setSurname(crimeApplyApplicant.getLastName());
-        applicant.setDateOfBirth(crimeApplyApplicant.getDateOfBirth());
-        applicant.setGender(APPLICANT_GENDER_DEFAULT_PREFER_NOT_TO_SAY);
-        applicant.setHasPartner(APPLICANT_HAS_PARTNER_DEFAULT_FALSE);
-        applicant.setTelephone(crimeApplyApplicant.getTelephoneNumber());
-        applicant.setNiNumber(crimeApplyApplicant.getNino());
-
-        uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.Applicant.CorrespondenceAddressType crimeApplyAddressType =
-                crimeApplyApplicant.getCorrespondenceAddressType();
-        applicant.setUseHomeAddress(mapUseHomeAddress(crimeApplyAddressType));
-        applicant.setNoFixedAbode(mapNoFixedAbode(crimeApplyApplicant.getHomeAddress()));
-        applicant.setUseSupplierAddressForPost(mapUseSupplierAddressForPost(crimeApplyAddressType));
-        applicant.setPartnerContraryInterest(mapPartnerContraryInterest());
-
-        applicant.setHomeAddress(mapAddress(crimeApplyApplicant.getHomeAddress()));
-        applicant.setPostalAddress(mapAddress(crimeApplyApplicant.getCorrespondenceAddress()));
-
-        return applicant;
-    }
-
-    private PartnerContraryInterest mapPartnerContraryInterest() {
-        PartnerContraryInterest partnerContraryInterest = new PartnerContraryInterest();
-        partnerContraryInterest.setCode(PARTNER_CONTRARY_INTEREST_CODE_DEFAULT_NOCON);
-        return partnerContraryInterest;
-    }
-
-    private boolean mapUseHomeAddress(uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.Applicant.
-                                              CorrespondenceAddressType crimeApplyAddressType) {
-
-        return uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.Applicant.
-                CorrespondenceAddressType.HOME_ADDRESS.equals(crimeApplyAddressType);
-    }
-
-    private boolean mapUseSupplierAddressForPost(uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.Applicant.CorrespondenceAddressType crimeApplyAddressType) {
-        if (crimeApplyAddressType == null) {
-            return COURT_CUSTODY_DEFAULT_FALSE;
-        }
-
-        switch (crimeApplyAddressType) {
-            case OTHER_ADDRESS, HOME_ADDRESS -> {
-                return COURT_CUSTODY_DEFAULT_FALSE;
-            }
-            case PROVIDERS_OFFICE_ADDRESS -> {
-                return true;
-            }
-            default -> throw new IllegalStateException("Unexpected value: " + crimeApplyAddressType);
-        }
-    }
-
-    private boolean mapNoFixedAbode(uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.general.Address crimeApplyAddress) {
-        return crimeApplyAddress == null;
-    }
-
-    private Address mapAddress(uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.general.Address crimeApplyAddress) {
-        if (crimeApplyAddress == null) {
-            return null;
-        }
-        Address address = new Address();
-        address.setLookupId(crimeApplyAddress.getLookupId());
-        address.setLine1(crimeApplyAddress.getAddressLineOne());
-        address.setLine2(crimeApplyAddress.getAddressLineTwo());
-        address.setCity(crimeApplyAddress.getCity());
-        address.setCountry(crimeApplyAddress.getCountry());
-        address.setPostCode(crimeApplyAddress.getPostcode());
-        return address;
     }
 }
